v2.5.1 pull request

+++ ita_root/ita_by_menu_export_import/backyard_main.py
@@ -179,7 +179,7 @@ def backyard_main(organization_id, workspace_id):
                 main_func_result, msg, trace_msg = menu_import_exec(objdbca, record, workspace_id, workspace_path, uploadfiles_dir, uploadfiles_60103_dir)
                 g.applogger.info("menu_import_exec END")
 
-            db_reconnention(objdbca, True) if objdbca else None
+            db_reconnention(objdbca) if objdbca else None
             # メイン処理がFalseの場合、異常系処理
             if not main_func_result:
                 # エラーログ出力
@@ -1485,7 +1485,6 @@ def menu_export_exec(objdbca, record, workspace_id, export_menu_dir, uploadfiles
         # ファイル名更新用パラメータを作成
         parameters = {
             "file": {
-                "file_name": file_encode(kym_path)
             },
             "parameter": {
                 "file_name": kym_name,
@@ -1494,7 +1493,13 @@ def menu_export_exec(objdbca, record, workspace_id, export_menu_dir, uploadfiles
             },
             "type": "Update"
         }
-
+        tmp_kym_path = kym_path.replace("/storage/", "/tmp/")
+        tmp_kym_dir_path = tmp_kym_path.replace(os.path.basename(kym_path), "")
+        os.makedirs(tmp_kym_dir_path, exist_ok=True)
+        shutil.move(kym_path, tmp_kym_dir_path)
+        record_file_paths = {
+            "file_name": tmp_kym_path
+        }
         # 「メニューエクスポート・インポート管理」のファイル名を更新
         objmenu = load_table.loadTable(objdbca, 'menu_export_import_list')
 
@@ -1503,7 +1508,7 @@ def menu_export_exec(objdbca, record, workspace_id, export_menu_dir, uploadfiles
         g.applogger.debug(debug_msg)
         objdbca.db_transaction_start()
 
-        exec_result = objmenu.exec_maintenance(parameters, execution_no, "", False, False, True, False, True)  # noqa: E999
+        exec_result = objmenu.exec_maintenance(parameters, execution_no, "", False, False, True, False, True, record_file_paths=record_file_paths)  # noqa: E999
 
         # コミット/トランザクション終了
         debug_msg = g.appmsg.get_log_message("BKY-20005", [])
@@ -1516,7 +1521,8 @@ def menu_export_exec(objdbca, record, workspace_id, export_menu_dir, uploadfiles
             raise Exception("499-00701", [result_msg])  # loadTableバリデーションエラー
 
         # 更新が正常に終了したらtmpの作業ファイルを削除する
-        os.remove(kym_path)
+        os.remove(kym_path) if os.path.isfile(kym_path) else None
+        os.remove(tmp_kym_path) if os.path.isfile(tmp_kym_path) else None
 
         # 正常系リターン
         return True, msg, None
@@ -1999,6 +2005,9 @@ def import_table_and_data(
     for record in menus_list:
         db_reconnention(objdbca) if objdbca else None
         db_reconnention(ws_db_sb) if ws_db_sb else None
+        db_reconnention(objdbca, True) if objdbca and not objdbca._is_transaction else None
+        db_reconnention(ws_db_sb, True) if ws_db_sb and not ws_db_sb._is_transaction  else None
+
         param = record['parameter']
         # menu_id = param.get('uuid')
         menu_id = param.get('menu_name')
@@ -2096,8 +2105,8 @@ def import_table_and_data(
                 else:
                     if objdbca._is_transaction is False and ita_base_menu_flg is True:
                         # トランザクション開始
-                        db_reconnention(objdbca, True) if objdbca else None
-                        db_reconnention(ws_db_sb, True) if ws_db_sb else None
+                        db_reconnention(objdbca) if objdbca else None
+                        db_reconnention(ws_db_sb) if ws_db_sb else None
                         debug_msg = g.appmsg.get_log_message("BKY-20004", [])
                         g.applogger.info(debug_msg)
                         objdbca.db_transaction_start()
@@ -2190,8 +2199,15 @@ def import_table_and_data(
 
         if objdbca._is_transaction is False and ita_base_menu_flg is False:
             # トランザクション開始
-            db_reconnention(objdbca, True) if objdbca else None
-            db_reconnention(ws_db_sb, True) if ws_db_sb else None
+            db_reconnention(objdbca) if objdbca else None
+            db_reconnention(ws_db_sb) if ws_db_sb else None
+            debug_msg = g.appmsg.get_log_message("BKY-20004", [])
+            g.applogger.info(debug_msg)
+            objdbca.db_transaction_start()
+        elif objdbca._is_transaction is False:
+            # トランザクション開始
+            db_reconnention(objdbca) if objdbca else None
+            db_reconnention(ws_db_sb) if ws_db_sb else None
             debug_msg = g.appmsg.get_log_message("BKY-20004", [])
             g.applogger.info(debug_msg)
             objdbca.db_transaction_start()
@@ -2229,7 +2245,6 @@ def import_table_and_data(
             debug_msg = g.appmsg.get_log_message("BKY-20005", [])
             g.applogger.info(debug_msg)
             objdbca.db_transaction_end(True)
-            objdbca._is_transaction = False
 
         # ファイル配置
         if file_put_flg and objmenu:
